requirements.txt
# Python 3.9+ required
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
google-generativeai>=0.8.0
httpx>=0.25.0
sse-starlette>=2.2.0
python-dotenv>=1.0.0